중복되는 클래스 제거하여 합침 전반적인 주석 추가, 수정 필요없는 코드 몇군데 삭제
