+++ src/main/java/com/tech/petfriends/helppetf/service/FindAddrTMapService.java
@@ -18,25 +18,25 @@ public class FindAddrTMapService implements HelppetfServiceInter {
 
 	@Override
 	public void execute(Model model) {
-		String apiKey = (String) model.getAttribute("apiKey");
-		model.addAttribute("apiKey", apiKey);
-		
 		model.addAttribute("main_navbar_id", "helppetf");
 		
+		// 모델에서 세션을 추출
 		HttpSession session = (HttpSession) model.getAttribute("session");
 		
+		// 세션에서 로그인 정보를 추출
 		MemberLoginDto memDto = (MemberLoginDto) session.getAttribute("loginUser");
 		
 		if (memDto == null) {
-			// 로그인이 안돼있을 경우 기본 주소를 서울시청으로
+			// 로그인 정보가 null인 경우 주소를 서울시청으로 지정
 			String userAddr = "서울특별시 중구 세종대로 110";
 			model.addAttribute("userAddr", userAddr);
 			model.addAttribute("mem_nick", null);
 		} else {
+			// 로그인 정보DTO에서 멤버 코드와 닉네임을 저장
 			String mem_nick = memDto.getMem_nick();
 			String mem_code = memDto.getMem_code();
-			model.addAttribute("mem_nick", mem_nick);
 			
+			// 멤버코드를 전달하여 멤버가 저장한 대표 주소를 반환받음
 			String userAddr = helpDao.findUserAddr(mem_code);
 			model.addAttribute("userAddr", userAddr);
 			model.addAttribute("mem_nick", mem_nick);
